Split context_menu.rs (534 lines) into context_menu/ subdirectory
Split crates/ui/src/context_menu.rs into focused modules:
- types.rs: ContextTarget, ContextMenuState, ContextMenuAction, PendingAction, zone_label helper
- detection.rs: detect_right_click_context_menu system
- ui_render.rs: context_menu_ui egui rendering system
- actions.rs: execute_context_menu_action system
- mod.rs: thin facade with plugin, mod declarations, and re-exports

All public items re-exported from mod.rs — no external import changes needed.
